Move intdiv into its own file
