+++ src/ModulusAlgorithm/DblAl.php
@@ -20,7 +20,7 @@ final class DblAl implements AlgorithmInterface
 
     public function quotient()
     {
-        return intdiv($this->result, 10);
+        return \Cs278\BankModulus\intdiv($this->result, 10);
     }
 
     public function remainder()
